Increase representability in delta
sqrt(1e-5) is a floating point representation of an irrational number between pow(2, -9) and pow(2, -8).
Floating point division with powers of 2 are done exactly. Division with roundings of irrational numbers have no such guarantee and have a very low likelyhood to be exact.
pow(2, -9) was chosen since it was a lower bound. Quantitative experiments should be done to see if pow(2, -8) is better. A version with pow(2, -8) is presented as a separate branch.
A version with representability control on sqrt(abs(y)U) is also available.