Significantly increased representability in delta
sqrt(1e-5) is a floating point representation of an irrational number between pow(2, -9) and pow(2, -8).
Floating point division with powers of 2 are done exactly. Division with roundings of irrational numbers have no such guarantee and have a very low likelyhood to be exact.
pow(2, -9) was chosen since it was a lower bound. Quantitative experiments should be done to see if pow(2, -8) is better.
This branch also improves on the representability by changing sqrt(|y|U) to the root of the magnitude of the step from |y| away from zero. For addition this makes no difference, but for division it guarantees representability.